refactor: rename postgres classes and remove breaklines
