+++ auto-configurations/models/chat/memory/spring-ai-autoconfigure-model-chat-memory-jdbc/src/test/java/org/springframework/ai/model/chat/memory/jdbc/autoconfigure/JdbcChatMemoryDataSourceScriptDatabasePostgresIT.java
@@ -17,10 +17,6 @@
 package org.springframework.ai.model.chat.memory.jdbc.autoconfigure;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.ai.chat.memory.jdbc.JdbcChatMemory;
-import org.springframework.ai.chat.messages.AssistantMessage;
-import org.springframework.ai.chat.messages.Message;
-import org.springframework.ai.chat.messages.UserMessage;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.JdbcTemplateAutoConfiguration;
@@ -30,8 +26,7 @@ import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
-import java.util.List;
-import java.util.UUID;
+import javax.sql.DataSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -39,14 +34,14 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @author Jonathan Leijendekker
  */
 @Testcontainers
-class JdbcChatMemoryAutoConfigurationPostgreSQLIT {
+class JdbcChatMemoryDataSourceScriptDatabasePostgresIT {
 
 	static final DockerImageName DEFAULT_IMAGE_NAME = DockerImageName.parse("postgres:17");
 
 	@Container
 	@SuppressWarnings("resource")
 	static PostgreSQLContainer<?> postgresContainer = new PostgreSQLContainer<>(DEFAULT_IMAGE_NAME)
-		.withDatabaseName("chat_memory_auto_configuration_test")
+		.withDatabaseName("chat_memory_initializer_test")
 		.withUsername("postgres")
 		.withPassword("postgres");
 
@@ -58,40 +53,13 @@ class JdbcChatMemoryAutoConfigurationPostgreSQLIT {
 				String.format("spring.datasource.password=%s", postgresContainer.getPassword()));
 
 	@Test
-	void jdbcChatMemoryScriptDatabaseInitializer_shouldBeLoaded() {
-		this.contextRunner.withPropertyValues("spring.ai.chat.memory.jdbc.initialize-schema=true")
-			.run(context -> assertThat(context.containsBean("jdbcChatMemoryScriptDatabaseInitializer")).isTrue());
-	}
+	void getSettings_shouldHaveSchemaLocations() {
+		this.contextRunner.run(context -> {
+			var dataSource = context.getBean(DataSource.class);
+			var settings = JdbcChatMemoryDataSourceScriptDatabaseInitializer.getSettings(dataSource);
 
-	@Test
-	void jdbcChatMemoryScriptDatabaseInitializer_shouldNotBeLoaded() {
-		this.contextRunner.withPropertyValues("spring.ai.chat.memory.jdbc.initialize-schema=false")
-			.run(context -> assertThat(context.containsBean("jdbcChatMemoryScriptDatabaseInitializer")).isFalse());
-	}
-
-	@Test
-	void addGetAndClear_shouldAllExecute() {
-		this.contextRunner.withPropertyValues("spring.ai.chat.memory.jdbc.initialize-schema=true").run(context -> {
-			var chatMemory = context.getBean(JdbcChatMemory.class);
-			var conversationId = UUID.randomUUID().toString();
-			var userMessage = new UserMessage("Message from the user");
-
-			chatMemory.add(conversationId, userMessage);
-
-			assertThat(chatMemory.get(conversationId, Integer.MAX_VALUE)).hasSize(1);
-			assertThat(chatMemory.get(conversationId, Integer.MAX_VALUE)).isEqualTo(List.of(userMessage));
-
-			chatMemory.clear(conversationId);
-
-			assertThat(chatMemory.get(conversationId, Integer.MAX_VALUE)).isEmpty();
-
-			var multipleMessages = List.<Message>of(new UserMessage("Message from the user 1"),
-					new AssistantMessage("Message from the assistant 1"));
-
-			chatMemory.add(conversationId, multipleMessages);
-
-			assertThat(chatMemory.get(conversationId, Integer.MAX_VALUE)).hasSize(multipleMessages.size());
-			assertThat(chatMemory.get(conversationId, Integer.MAX_VALUE)).isEqualTo(multipleMessages);
+			assertThat(settings.getSchemaLocations())
+				.containsOnly("classpath:org/springframework/ai/chat/memory/jdbc/schema-postgresql.sql");
 		});
 	}
