+++ memory/spring-ai-model-chat-memory-jdbc/src/test/java/org/springframework/ai/chat/memory/jdbc/JdbcChatMemoryPostgresIT.java
@@ -49,7 +49,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
  * @author Xavier Chopin
  */
 @Testcontainers
-class JdbcChatMemoryPostgreSQLIT {
+class JdbcChatMemoryPostgresIT {
 
 	@Container
 	@SuppressWarnings("resource")
